more PR fixes
