+++ api/python/cell_census/src/cell_census/__init__.py
@@ -1,8 +1,8 @@
+from .experiment import get_experiment
 from .get_anndata import get_anndata
-from .get_presence_matrix import get_presence_matrix
 from .open import download_source_h5ad, get_source_h5ad_uri, open_soma
+from .presence_matrix import get_presence_matrix
 from .release_directory import get_census_version_description, get_census_version_directory
-from .util import get_experiment
 
 __version__ = "0.0.1-dev0"
